Resuming works now
git-svn-id: http://cyanogen-updater.googlecode.com/svn/trunk@574 fa1c2e9a-81d2-11de-92f0-6bbadbc12360

+++ src/cmupdaterapp/customTypes/DownloadProgress.java
@@ -2,13 +2,13 @@ package cmupdaterapp.customTypes;
 
 public class DownloadProgress
 {
-	private final int downloaded;
+	private final long downloaded;
 	private final int total;
 	private final String downloadedText;
 	private final String speedText;
 	private final String remainingTimeText;
 	
-	public DownloadProgress(int _downloaded, int _total, String _downloadedText, String _speedText, String _remainingTimeText)
+	public DownloadProgress(long _downloaded, int _total, String _downloadedText, String _speedText, String _remainingTimeText)
 	{
 		downloaded = _downloaded;
 		total = _total;
@@ -17,7 +17,7 @@ public class DownloadProgress
 		remainingTimeText = _remainingTimeText;
 	}
 	
-	public int getDownloaded() { return downloaded; }
+	public long getDownloaded() { return downloaded; }
 	public int getTotal() { return total; }
 	public String getDownloadedText() { return downloadedText; }
 	public String getSpeedText() { return speedText; }
